+++ addOns/spider/src/main/java/org/zaproxy/addon/spider/SpiderParam.java
@@ -103,6 +103,8 @@ public class SpiderParam extends VersionedAbstractParam {
     /** The Constant SPIDER_HANDLE_ODATA_PARAMETERS. */
     private static final String SPIDER_HANDLE_ODATA_PARAMETERS = "spider.handleODataParameters";
 
+    private static final String LOGOUT_AVOIDANCE = "spider.logoutAvoidance";
+
     private static final String DOMAIN_ALWAYS_IN_SCOPE_KEY = "spider.domainsAlwaysInScope";
     private static final String ALL_DOMAINS_ALWAYS_IN_SCOPE_KEY =
             DOMAIN_ALWAYS_IN_SCOPE_KEY + ".domainAlwaysInScope";
@@ -151,6 +153,8 @@ public class SpiderParam extends VersionedAbstractParam {
     private ExtensionHttpSessions extensionHttpSessions;
 
     private ExtensionAntiCSRF extensionAntiCSRF;
+
+    public static final boolean DEFAULT_LOGOUT_AVOIDANCE = false;
 
     /**
      * This option is used to define how the parameters are used when checking if an URI was already
@@ -231,6 +235,8 @@ public class SpiderParam extends VersionedAbstractParam {
      * known URL *
      */
     private boolean handleODataParametersVisited;
+
+    private boolean logoutAvoidance;
 
     /** The maximum duration in minutes that the spider is allowed to run for, 0 meaning no limit */
     private int maxDuration;
@@ -367,6 +373,8 @@ public class SpiderParam extends VersionedAbstractParam {
         loadIrrelevantParameters();
         this.confirmRemoveIrrelevantParameter =
                 getBoolean(CONFIRM_REMOVE_IRRELEVANT_PARAMETER, true);
+
+        this.logoutAvoidance = getBoolean(LOGOUT_AVOIDANCE, DEFAULT_LOGOUT_AVOIDANCE);
     }
 
     @Override
@@ -1166,5 +1174,14 @@ public class SpiderParam extends VersionedAbstractParam {
         this.confirmRemoveIrrelevantParameter = confirmRemove;
         getConfig()
                 .setProperty(CONFIRM_REMOVE_IRRELEVANT_PARAMETER, confirmRemoveIrrelevantParameter);
+    }
+
+    public boolean isLogoutAvoidance() {
+        return logoutAvoidance;
+    }
+
+    public void setLogoutAvoidance(boolean avoidLogout) {
+        this.logoutAvoidance = avoidLogout;
+        getConfig().setProperty(LOGOUT_AVOIDANCE, Boolean.toString(avoidLogout));
     }
 }